Move out vg compare
