+++ src/main.cpp
@@ -38,131 +38,6 @@ using namespace std;
 using namespace google::protobuf;
 using namespace vg;
 
-void help_compare(char** argv) {
-    cerr << "usage: " << argv[0] << " compare [options] graph1 graph2" << endl
-        << "Compare kmer sets of two graphs" << endl
-        << endl
-        << "options:" << endl
-        << "    -d, --db-name1 FILE  use this db for graph1 (defaults to <graph1>.index/)" << endl
-        << "    -e, --db-name2 FILE  use this db for graph2 (defaults to <graph1>.index/)" << endl
-        << "    -t, --threads N      number of threads to use" << endl;
-}
-
-int main_compare(int argc, char** argv) {
-
-    if (argc <= 3) {
-        help_compare(argv);
-        return 1;
-    }
-
-    string db_name1;
-    string db_name2;
-    int num_threads = 1;
-
-    int c;
-    optind = 2; // force optind past command positional argument
-    while (true) {
-        static struct option long_options[] =
-        {
-            {"help", no_argument, 0, 'h'},
-            {"db-name1", required_argument, 0, 'd'},
-            {"db-name2", required_argument, 0, 'e'},
-            {"threads", required_argument, 0, 't'},
-            {0, 0, 0, 0}
-        };
-
-        int option_index = 0;
-        c = getopt_long (argc, argv, "hd:e:t:",
-                long_options, &option_index);
-
-        // Detect the end of the options.
-        if (c == -1)
-            break;
-
-        switch (c)
-        {
-
-            case 'd':
-                db_name1 = optarg;
-                break;
-
-            case 'e':
-                db_name2 = optarg;
-                break;
-
-            case 't':
-                num_threads = atoi(optarg);
-                break;
-
-            case 'h':
-            case '?':
-                help_compare(argv);
-                exit(1);
-                break;
-
-            default:
-                abort ();
-        }
-    }
-
-    omp_set_num_threads(num_threads);
-
-    if (db_name1.empty()) {
-        db_name1 = get_input_file_name(optind, argc, argv);
-    }
-    if (db_name2.empty()) {
-        db_name2 = get_input_file_name(optind, argc, argv);
-    }
-
-    // Note: only supporting rocksdb index for now.
-
-    Index index1;
-    index1.open_read_only(db_name1);
-
-    Index index2;
-    index2.open_read_only(db_name2);
-
-    pair<int64_t, int64_t> index1_vs_index2;
-    pair<int64_t, int64_t> index2_vs_index1;
-
-    // Index::compare is not parallel, but at least we can do the
-    // two directions at the same time...
-#pragma omp parallel sections
-    {
-#pragma omp section
-        {
-            index1_vs_index2 = index1.compare_kmers(index2);
-        }
-#pragma omp section
-        {
-            index2_vs_index1 = index2.compare_kmers(index1);
-        }
-    }
-    {// <-- for emacs
-        assert(index1_vs_index2.first == index2_vs_index1.first);
-
-        int64_t db1_count = index1_vs_index2.first + index1_vs_index2.second;
-        int64_t db2_count = index2_vs_index1.first + index2_vs_index1.second;
-        int64_t db1_only = index1_vs_index2.second;
-        int64_t db2_only = index2_vs_index1.second;
-        int64_t db1_and_db2 = index1_vs_index2.first;
-        int64_t db1_or_db2 = db1_only + db2_only + db1_and_db2;
-
-        cout << "{\n"
-            << "\"db1_path\": " << "\"" << db_name1 << "\"" << ",\n"
-            << "\"db2_path\": " << "\"" << db_name2 << "\"" << ",\n"
-            << "\"db1_total\": " << db1_count << ",\n"
-            << "\"db2_total\": " << db2_count << ",\n"
-            << "\"db1_only\": " << db1_only << ",\n"
-            << "\"db2_only\": " << db2_only << ",\n"
-            << "\"intersection\": " << db1_and_db2 << ",\n"
-            << "\"union\": " << db1_or_db2 << "\n"
-            << "}" << endl;
-    }
-    return 0;
-}
-
-
 void help_pileup(char** argv) {
     cerr << "usage: " << argv[0] << " pileup [options] <graph.vg> <alignment.gam> > out.vgpu" << endl
          << "Calculate pileup for each position in graph and output in VG Pileup format (list of protobuf NodePileups)." << endl
@@ -1617,7 +1492,6 @@ void vg_help(char** argv) {
          << "  -- kmers         enumerate kmers of the graph" << endl
          << "  -- pileup        build a pileup from a set of alignments" << endl
          << "  -- genotype      compute genotypes from aligned reads" << endl
-         << "  -- compare       compare the kmer space of two graphs" << endl
          << "  -- circularize   circularize a path within a graph." << endl
          << "  -- sort          sort variant graph using max flow algorithm or Eades fast heuristic algorithm" << endl
          << "  -- test          run unit tests" << endl;
@@ -1659,8 +1533,6 @@ int main(int argc, char *argv[])
         return main_kmers(argc, argv);
     } else if (command == "pileup") {
         return main_pileup(argc, argv);
-    } else if (command == "compare") {
-        return main_compare(argc, argv);
     } else if (command == "circularize"){
         return main_circularize(argc, argv);
     } else if (command == "test") {
